requirements.txt
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
drf_yasg==1.21.6